+++ src/Recyclarr/Config/IConfigurationLoader.cs
@@ -2,8 +2,10 @@ using TrashLib.Config.Services;
 
 namespace Recyclarr.Config;
 
-public interface IConfigurationLoader<out T>
+public interface IConfigurationLoader<T>
     where T : IServiceConfiguration
 {
-    IEnumerable<T> LoadMany(IEnumerable<string> configFiles, string configSection);
+    ICollection<T> LoadMany(IEnumerable<string> configFiles, string configSection);
+    ICollection<T> Load(string file, string configSection);
+    ICollection<T> LoadFromStream(TextReader stream, string requestedSection);
 }