+++ src/Recyclarr/Logging/LogProperty.cs
@@ -1,0 +1,6 @@
+namespace Recyclarr.Logging;
+
+public static class LogProperty
+{
+    public static string Scope => nameof(Scope);
+}