+++ src/TrashLib/Config/IYamlSerializerFactory.cs
@@ -4,6 +4,6 @@ namespace TrashLib.Config;
 
 public interface IYamlSerializerFactory
 {
-    IDeserializer CreateDeserializer();
+    IDeserializer CreateDeserializer(Action<DeserializerBuilder>? extraBuilder = null);
     ISerializer CreateSerializer();
 }