+++ src/TrashLib/Config/Secrets/SecretNotFoundException.cs
@@ -1,0 +1,14 @@
+namespace TrashLib.Config.Secrets;
+
+public class SecretNotFoundException : Exception
+{
+    public int Line { get; }
+    public string SecretKey { get; }
+
+    public SecretNotFoundException(int line, string secretKey)
+        : base($"Secret used on line {line} with key {secretKey} is not defined in secrets.yml")
+    {
+        Line = line;
+        SecretKey = secretKey;
+    }
+}